+++ audio-type.js
