+++ package.json
@@ -5,23 +5,23 @@
   "license": "MIT",
   "repository": "audiojs/audio-type",
   "type": "module",
+  "main": "audio-type.js",
   "author": {
     "name": "Hemanth.HM",
     "email": "hemanth.hm@gmail.com",
     "url": "http://h3manth.com"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=14"
   },
   "scripts": {
-    "test": "node test",
-    "browser": "browserify -s $npm_package_name -o browser.js ."
+    "test": "node test"
   },
   "files": [
-    "index.js",
-    "cli.js"
+    "audio-type.js"
   ],
   "keywords": [
+    "audiojs",
     "audio",
     "sound",
     "wav",
@@ -32,9 +32,7 @@
     "check",
     "is",
     "binary",
-    "buffer",
-    "uint8array",
-    "cli"
+    "uint8array"
   ],
   "devDependencies": {
     "read-chunk": "^2.0.0",
